Chore: Refactor mobile plugin loading logic into more proper locations
