+++ .gitignore
@@ -575,6 +575,27 @@ packages/app-mobile/components/buttons/index.js
 packages/app-mobile/components/getResponsiveValue.test.js
 packages/app-mobile/components/getResponsiveValue.js
 packages/app-mobile/components/global-style.js
+packages/app-mobile/components/plugins/PluginRunner.js
+packages/app-mobile/components/plugins/PluginRunnerWebView.js
+packages/app-mobile/components/plugins/backgroundPage/initializeDialogWebView.js
+packages/app-mobile/components/plugins/backgroundPage/initializePluginBackgroundIframe.js
+packages/app-mobile/components/plugins/backgroundPage/pluginRunnerBackgroundPage.js
+packages/app-mobile/components/plugins/backgroundPage/startStopPlugin.js
+packages/app-mobile/components/plugins/backgroundPage/utils/getFormData.test.js
+packages/app-mobile/components/plugins/backgroundPage/utils/getFormData.js
+packages/app-mobile/components/plugins/backgroundPage/utils/makeSandboxedIframe.js
+packages/app-mobile/components/plugins/backgroundPage/utils/reportUnhandledErrors.js
+packages/app-mobile/components/plugins/backgroundPage/utils/wrapConsoleLog.js
+packages/app-mobile/components/plugins/dialogs/PluginDialogManager.js
+packages/app-mobile/components/plugins/dialogs/PluginDialogWebView.js
+packages/app-mobile/components/plugins/dialogs/PluginPanelViewer.js
+packages/app-mobile/components/plugins/dialogs/PluginUserWebView.js
+packages/app-mobile/components/plugins/dialogs/hooks/useDialogMessenger.js
+packages/app-mobile/components/plugins/dialogs/hooks/useDialogSize.js
+packages/app-mobile/components/plugins/dialogs/hooks/useViewInfos.js
+packages/app-mobile/components/plugins/dialogs/hooks/useWebViewSetup.js
+packages/app-mobile/components/plugins/types.js
+packages/app-mobile/components/plugins/utils/createOnLogHandler.js
 packages/app-mobile/components/screens/ConfigScreen/ConfigScreen.js
 packages/app-mobile/components/screens/ConfigScreen/FileSystemPathSelector.js
 packages/app-mobile/components/screens/ConfigScreen/JoplinCloudConfig.js
@@ -636,33 +657,11 @@ packages/app-mobile/components/screens/status.js
 packages/app-mobile/components/side-menu-content.js
 packages/app-mobile/components/voiceTyping/VoiceTypingDialog.js
 packages/app-mobile/gulpfile.js
-packages/app-mobile/plugins/PlatformImplementation.js
-packages/app-mobile/plugins/PluginRunner/PluginRunner.js
-packages/app-mobile/plugins/PluginRunner/PluginRunnerWebView.js
-packages/app-mobile/plugins/PluginRunner/backgroundPage/initializeDialogWebView.js
-packages/app-mobile/plugins/PluginRunner/backgroundPage/initializePluginBackgroundIframe.js
-packages/app-mobile/plugins/PluginRunner/backgroundPage/pluginRunnerBackgroundPage.js
-packages/app-mobile/plugins/PluginRunner/backgroundPage/startStopPlugin.js
-packages/app-mobile/plugins/PluginRunner/backgroundPage/utils/getFormData.test.js
-packages/app-mobile/plugins/PluginRunner/backgroundPage/utils/getFormData.js
-packages/app-mobile/plugins/PluginRunner/backgroundPage/utils/makeSandboxedIframe.js
-packages/app-mobile/plugins/PluginRunner/backgroundPage/utils/reportUnhandledErrors.js
-packages/app-mobile/plugins/PluginRunner/backgroundPage/utils/wrapConsoleLog.js
-packages/app-mobile/plugins/PluginRunner/dialogs/PluginDialogManager.js
-packages/app-mobile/plugins/PluginRunner/dialogs/PluginDialogWebView.js
-packages/app-mobile/plugins/PluginRunner/dialogs/PluginPanelViewer.js
-packages/app-mobile/plugins/PluginRunner/dialogs/PluginUserWebView.js
-packages/app-mobile/plugins/PluginRunner/dialogs/hooks/useDialogMessenger.js
-packages/app-mobile/plugins/PluginRunner/dialogs/hooks/useDialogSize.js
-packages/app-mobile/plugins/PluginRunner/dialogs/hooks/useViewInfos.js
-packages/app-mobile/plugins/PluginRunner/dialogs/hooks/useWebViewSetup.js
-packages/app-mobile/plugins/PluginRunner/types.js
-packages/app-mobile/plugins/PluginRunner/utils/createOnLogHandler.js
-packages/app-mobile/plugins/hooks/usePlugin.js
 packages/app-mobile/root.js
 packages/app-mobile/services/AlarmServiceDriver.android.js
 packages/app-mobile/services/AlarmServiceDriver.ios.js
 packages/app-mobile/services/e2ee/RSA.react-native.js
+packages/app-mobile/services/plugins/PlatformImplementation.js
 packages/app-mobile/services/profiles/index.js
 packages/app-mobile/services/voiceTyping/vosk.android.js
 packages/app-mobile/services/voiceTyping/vosk.ios.js
@@ -871,6 +870,7 @@ packages/lib/geolocation-node.js
 packages/lib/hooks/useAsyncEffect.js
 packages/lib/hooks/useElementSize.js
 packages/lib/hooks/useEventListener.js
+packages/lib/hooks/usePlugin.js
 packages/lib/hooks/usePrevious.js
 packages/lib/htmlUtils.test.js
 packages/lib/htmlUtils.js
