+++ packages/app-mobile/components/plugins/backgroundPage/initializeDialogWebView.ts
