+++ packages/app-mobile/components/plugins/backgroundPage/initializePluginBackgroundIframe.ts
