+++ packages/app-mobile/components/plugins/backgroundPage/startStopPlugin.ts
