+++ packages/app-mobile/components/plugins/backgroundPage/utils/getFormData.ts
