+++ packages/app-mobile/components/plugins/backgroundPage/utils/makeSandboxedIframe.ts
