+++ packages/app-mobile/components/plugins/backgroundPage/utils/reportUnhandledErrors.ts
