+++ packages/app-mobile/components/plugins/backgroundPage/utils/wrapConsoleLog.ts
