+++ packages/app-mobile/components/plugins/dialogs/PluginPanelViewer.tsx
