+++ packages/app-mobile/components/plugins/dialogs/hooks/useViewInfos.ts
