+++ packages/app-mobile/components/plugins/utils/createOnLogHandler.ts
