+++ packages/app-mobile/components/screens/ConfigScreen/plugins/utils/usePluginItem.ts
@@ -2,7 +2,7 @@ import { PluginItem } from '@joplin/lib/components/shared/config/plugins/types';
 import { PluginSettings } from '@joplin/lib/services/plugins/PluginService';
 import { PluginManifest } from '@joplin/lib/services/plugins/utils/types';
 import { useMemo, useRef } from 'react';
-import usePlugin from '../../../../../plugins/hooks/usePlugin';
+import usePlugin from '@joplin/lib/hooks/usePlugin';
 
 // initialItem is used when the plugin is not installed. For example, if the plugin item is being
 // created from search results.
