+++ packages/lib/hooks/usePlugin.ts
@@ -1,13 +1,14 @@
-import PluginService from '@joplin/lib/services/plugins/PluginService';
+import PluginService from '../services/plugins/PluginService';
 import Logger from '@joplin/utils/Logger';
-import { useEffect, useMemo, useRef, useState } from 'react';
+import shim from '../shim';
 
 const logger = Logger.create('usePlugin');
 
 const usePlugin = (pluginId: string) => {
-	const [pluginReloadCounter, setPluginReloadCounter] = useState(0);
+	const React = shim.react();
+	const [pluginReloadCounter, setPluginReloadCounter] = React.useState(0);
 
-	const plugin = useMemo(() => {
+	const plugin = React.useMemo(() => {
 		if (!PluginService.instance().pluginIds.includes(pluginId)) {
 			return null;
 		}
@@ -21,12 +22,12 @@ const usePlugin = (pluginId: string) => {
 		// matches the one loaded in the PluginService.
 	}, [pluginId, pluginReloadCounter]);
 
-	const reloadCounterRef = useRef(0);
+	const reloadCounterRef = React.useRef(0);
 	reloadCounterRef.current = pluginReloadCounter;
 
 	// The plugin may need to be re-fetched from the PluginService. When a plugin is reloaded,
 	// its Plugin object is replaced with a new one.
-	useEffect(() => {
+	React.useEffect(() => {
 		const { remove } = PluginService.instance().addLoadedPluginsChangeListener(() => {
 			setPluginReloadCounter(reloadCounterRef.current + 1);
 		});
